rembgがいないバージョンのAPIがコードを汚していたので,まとめてno_rembgに記述.
