+++ api/chat_gpt/car_data.py
@@ -27,22 +27,3 @@ async def make_car(player: str,text: str, api_key: str = Security(validate_api_k
 
 
     return {"url_car_img": url_car_img,"name": name,"luk": luk,"text_car_status": text_jp}
-'''
-
-@router.get("/{player}/car/data")
-async def make_car(player: str,text: str, api_key: str = Security(validate_api_key)):
-
-    text_en = translation(text,'JA','EN-US')
-    
-    if validate_token_count(text_en,5):
-        url_car_img, [luk,name,text_car_status] = await asyncio.gather(
-            image_generate_chatgpt_no_rembg(text_en),
-            status_generate_chatgpt(text_en)
-        )
-
-    
-    text_jp = translation(text_car_status,'EN','JA')
-
-
-    return {"url_car_img": url_car_img,"name": name,"luk": luk,"text_car_status": text_jp}
-'''
