+++ api/chat_gpt/image_generation.py
@@ -40,27 +40,5 @@ async def image_generate_chatgpt(text:str):
     binary_image = buffered.getvalue() # 画像(バイナリ)を取得
     
     return b64encode(binary_image) # 画像(バイナリ)をbase64に変換して返す
-"""
 
-async def image_generate_chatgpt_no_rembg(text:str):
-    
-    text_prompt = shaping_prompts_car_img(text)
-    #modelはdell-e2
-    response =  client.images.generate(
-                        model   = "dall-e-2",   # モデル  
-                        prompt  = text_prompt,         # 画像生成に用いる説明文章         
-                        n       = 1,            # 何枚の画像を生成するか  
-                        #size="1024x1024",
-                        size="256x256",
-                        quality="standard",
-                    )
-    
-    image_url = response.data[0].url
-
-    # URLから画像(バイナリ)を取得
-    car_img_binary = requests.get(image_url).content
-    
-    
-    return b64encode(car_img_binary) # 画像(バイナリ)をbase64に変換して返す
-"""
